+++ CHANGES.rst
@@ -21,6 +21,9 @@ Changelog
   when using `collective_contact_core_workflow`, an element in state `deactivated` is no more viewable by `Member`
 - Manage missing terms for `SelectableAssemblyMembersVocabulary` and `SelectableItemInitiatorsVocabulary` as now, inactive `held_position` objects
   are no more returned by default by these vocabularies (only `active` elements are returned)
+- Renamed `Products.PloneMeeting.vocabularies.selectableassociatedorganizationsvocabulary` to
+  `Products.PloneMeeting.vocabularies.detailedorganizationsvocabulary` so it is easier to reuse in other contexts
+- Added possibility to select organizations as item initiators (`MeetingItem.itemInitiator`) in addition to held positions
 
 4.1.23.3 (2020-04-30)
 ---------------------